+++ generate-docs.sh
@@ -17,7 +17,9 @@ echo "Generating documentation"
 doxygen docs/doxy
 
 # fix broken links to SCIP online documentation
-sed -i "s/\.php\.html/\.php/g" docs/html/*.html
+# If you set `HTML_FILE_EXTENSION    = .php` in doc/doxy you don't need the following sed commands
+sed -i "s@\.php\.html@.php@g" docs/html/*.* docs/html/search/*.*
+sed -i -E "s@(scip.zib.de.*)\.html@\1.php@g" docs/html/*.* docs/html/search/*.*
 
 # clone the docu branch
 git clone -b gh-pages git@github.com:${GH_REPO_ORG}/${GH_REPO_NAME} code_docs